integration-test: move DeleteCommandTest out from LogicManagerTest
The LogicManagerTest tests all commands in integration test level.

It's very heavy and hard to maintain.

Let's break it down. This step moves out tests for the delete command.
