+++ src/test/java/seedu/address/integration/DeleteCommandTest.java
@@ -1,0 +1,38 @@
+package seedu.address.integration;
+
+import static seedu.address.commons.core.Messages.MESSAGE_INVALID_COMMAND_FORMAT;
+
+import java.util.List;
+
+import org.junit.Test;
+
+import seedu.address.logic.commands.DeleteCommand;
+import seedu.address.model.AddressBook;
+import seedu.address.model.person.Person;
+
+public class DeleteCommandTest extends AddressBookIntegrationTest {
+
+    @Test
+    public void execute_deleteInvalidArgsFormat_errorMessageShown() throws Exception {
+        String expectedMessage = String.format(MESSAGE_INVALID_COMMAND_FORMAT, DeleteCommand.MESSAGE_USAGE);
+        assertIncorrectIndexFormatBehaviorForCommand("delete", expectedMessage);
+    }
+
+    @Test
+    public void execute_deleteIndexNotFound_errorMessageShown() throws Exception {
+        assertIndexNotFoundBehaviorForCommand("delete");
+    }
+
+    @Test
+    public void execute_delete_removesCorrectPerson() throws Exception {
+        TestDataHelper helper = new TestDataHelper();
+        List<Person> threePersons = helper.generatePersonList(3);
+        AddressBook expectedAB = helper.generateAddressBook(threePersons);
+        expectedAB.removePerson(threePersons.get(1));
+        helper.addToModel(model, threePersons);
+        assertCommandSuccess("delete 2",
+                String.format(DeleteCommand.MESSAGE_DELETE_PERSON_SUCCESS, threePersons.get(1)),
+                expectedAB,
+                expectedAB.getPersonList());
+    }
+}
